+++ Parser.h
@@ -18,7 +18,7 @@ class Parser {
   unsigned int sequencePosition;
   // The file to write to.
   std::string filename;
-  unsigned long long epochToMidnightLocalNanos;
+  uint64_t epochToMidnightLocalNanos;
 
   // Payload bytes that have been received but not yet processed.
   std::queue<char> q;
@@ -29,9 +29,9 @@ class Parser {
   std::unordered_map<uint16_t, const char*> packets;
 
   // Track Add Orders and their remaining order size.
-  std::unordered_map<unsigned long long, Order_t> orders;
+  std::unordered_map<uint64_t, Order_t> orders;
   // Utility to abstract away unordered_map boilerplate, lacking ::contains.
-  Order_t* lookupOrder(unsigned long long orderRef);
+  Order_t* lookupOrder(uint64_t orderRef);
 
   // Functions that map message from input to output message in second buffer.
   void mapAdd(const char* in, char** outPtr);
